backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
langchain==1.3.0
langchain-upstage==0.7.7
pillow==11.2.1
PyMuPDF==1.27.2.3
python-dotenv==1.0.1
